Abuse module constant space for fun and profit
This module takes an arbitrary list of tuples, and a
pair of functions, one to set a key and a value for
each element of the tuple list.

It compiles these into a module which is then loaded
into an Erlang system which provides off-heap shared
access without copying terms into each process.

The primary way to look up information is to use
get/2 or get/3.